requirements.txt
pymobiledevice3==5.1.2
python-socketio==5.14.3
tenacity==9.1.2
uvicorn==0.38.0
